fixed #832 - display health as int

+++ src/main/java/de/raidcraft/skills/tabdeco/TabDecoMaxHealthSettings.java
@@ -19,6 +19,6 @@ public class TabDecoMaxHealthSettings extends TabDecoSetting {
     @Override
     public String getSlotText(Player player, String inputText, String settingName) {
 
-        return String.valueOf(player.getMaxHealth());
+        return String.valueOf((int) player.getMaxHealth());
     }
 }
